[1.21.5] More template fixes woo

+++ src/main/java/com/bergerkiller/reflection/net/minecraft/server/NMSPacketClasses.java
@@ -695,8 +695,8 @@ public class NMSPacketClasses {
     public static class NMSPacketPlayInWindowClick extends NMSPacket {
 
         public final FieldAccessor<Integer> windowId = PacketPlayInWindowClickHandle.T.windowId.toFieldAccessor();
-        public final FieldAccessor<Integer> slot = PacketPlayInWindowClickHandle.T.slot.toFieldAccessor();
-        public final FieldAccessor<Integer> button = PacketPlayInWindowClickHandle.T.button.toFieldAccessor();
+        public final FieldAccessor<Short> slot = PacketPlayInWindowClickHandle.T.slot.toFieldAccessor();
+        public final FieldAccessor<Byte> button = PacketPlayInWindowClickHandle.T.button.toFieldAccessor();
         public final FieldAccessor<ItemStack> item = PacketPlayInWindowClickHandle.T.item.toFieldAccessor();
         public final FieldAccessor<InventoryClickType> mode = PacketPlayInWindowClickHandle.T.mode.toFieldAccessor();
     }
